Factor actions.groups out of actions.layers to tame the size of the latter module.
