+++ src/js/actions/mask.js
@@ -24,25 +24,115 @@
 define(function (require, exports) {
     "use strict";
 
-    var Promise = require("bluebird");
-
-    var layerActions = require("./layers"),
-        menuActions = require("./menu"),
-        toolActions = require("./tools"),
-        locks = require("../locks");
+    var Promise = require("bluebird"),
+        Immutable = require("immutable");
 
     var descriptor = require("adapter").ps.descriptor,
         boundsLib = require("adapter").lib.bounds,
         vectorMaskLib = require("adapter").lib.vectorMask,
         documentLib = require("adapter").lib.document;
 
-    var nls = require("js/util/nls");
+    var locks = require("../locks"),
+        events = require("../events"),
+        nls = require("js/util/nls");
 
     var _CLEAR_PATH = 106;
 
     /**
-     * Handle escaping from vector mask for the selection based vector mask tools
-     * most tools can just call toolActions.changeVectorMaskMode directly
+     * Select the vector mask for the currently selected layer.
+     *
+     * @return {Promise}
+     */
+    var selectVectorMask = function () {
+        return descriptor.playObject(vectorMaskLib.selectVectorMask());
+    };
+
+    selectVectorMask.action = {
+        reads: [],
+        writes: [locks.PS_DOC],
+        modal: true
+    };
+
+    /**
+     * Reveal and select the vector mask of the selected layer. 
+     * Also switch to the vector based superselect tool.
+     *
+     * @return {Promise}
+     */
+    var editVectorMask = function () {
+        var toolStore = this.flux.store("tool"),
+            superselectVector = toolStore.getToolByID("superselectVector");
+
+        return this.transfer("tools.select", superselectVector)
+            .then(function () {
+                // select and activate knots on Current Vector Mask
+                return descriptor.playObject(vectorMaskLib.activateVectorMaskEditing());
+            });
+    };
+    editVectorMask.action = {
+        reads: [locks.JS_TOOL],
+        writes: [locks.PS_DOC],
+        transfers: ["tools.select"],
+        modal: true
+    };
+
+    /**
+     * Removes a vector mask on the selected layer.
+     *
+     * @return {Promise}
+     */
+    var deleteVectorMask = function () {
+        var applicationStore = this.flux.store("application"),
+            currentDocument = applicationStore.getCurrentDocument();
+
+        if (currentDocument === null) {
+            return Promise.resolve();
+        }
+
+        var layers = currentDocument.layers.selected;
+
+        if (layers === null || layers.isEmpty()) {
+            return Promise.resolve();
+        }
+
+        var currentLayer = layers.first();
+
+        if (currentLayer.vectorMaskEnabled) {
+            var deleteMaskOptions = {
+                historyStateInfo: {
+                    name: nls.localize("strings.ACTIONS.DELETE_VECTOR_MASK"),
+                    target: documentLib.referenceBy.id(currentDocument.id)
+                }
+            };
+            return descriptor.playObject(vectorMaskLib.deleteVectorMask(), deleteMaskOptions)
+                .bind(this)
+                .then(function () {
+                    var payload = {
+                            documentID: currentDocument.id,
+                            layerIDs: Immutable.List.of(currentLayer.id),
+                            vectorMaskEnabled: false,
+                            history: {
+                                newState: true,
+                                name: nls.localize("strings.ACTIONS.DELETE_VECTOR_MASK")
+                            }
+                        },
+                        event = events.document.history.REMOVE_VECTOR_MASK_FROM_LAYER;
+
+                    return this.dispatchAsync(event, payload);
+                });
+        } else {
+            return Promise.resolve();
+        }
+    };
+    deleteVectorMask.action = {
+        reads: [locks.JS_TOOL],
+        writes: [locks.PS_DOC, locks.JS_DOC]
+    };
+    
+
+    /**
+     * Handle escaping from vector mask for the selection-based vector mask tools.
+     * Most tools can just call tools.changeVectorMaskMode directly.
      *
      * @return {Promise}
      */
@@ -56,7 +146,7 @@ define(function (require, exports) {
         }
 
         if (toolStore.getVectorMode() && toolStore.getModalToolState()) {
-            return this.transfer(toolActions.changeVectorMaskMode, false);
+            return this.transfer("tools.changeVectorMaskMode", false);
         }
 
         var currentLayers = currentDocument.layers.selected;
@@ -67,27 +157,27 @@ define(function (require, exports) {
 
         var currentLayer = currentLayers.first();
 
-        return this.transfer(layerActions.resetLayers, currentDocument, currentLayer)
+        return this.transfer("layers.resetLayers", currentDocument, currentLayer)
             .bind(this)
             .then(function () {
                 currentDocument = appStore.getCurrentDocument();
                 currentLayer = currentDocument.layers.selected.first();
                 if (currentLayer && currentLayer.vectorMaskEmpty) {
-                    return this.transfer(toolActions.changeVectorMaskMode, false);
+                    return this.transfer("tools.changeVectorMaskMode", false);
                 } else {
-                    return this.transfer(toolActions.select, toolStore.getToolByID("newSelect"));
+                    return this.transfer("tools.select", toolStore.getToolByID("newSelect"));
                 }
             });
     };
     handleEscapeVectorMask.action = {
         reads: [locks.JS_APP, locks.JS_DOC],
         writes: [],
-        transfers: [layerActions.resetLayers, toolActions.changeVectorMaskMode, toolActions.select],
+        transfers: ["layers.resetLayers", "tools.changeVectorMaskMode", "tools.select"],
         modal: true
     };
 
     /**
-     * Handle deleting vector mask for all of the vector mask mode tools
+     * Handle deleting the vector mask for all of the vector mask mode tools.
      *
      * @return {Promise}
      */
@@ -101,16 +191,16 @@ define(function (require, exports) {
         }
 
         if (toolStore.getModalToolState()) {
-            return this.transfer(layerActions.deleteVectorMask)
+            return this.transfer("mask.deleteVectorMask")
                 .bind(this)
                 .then(function () {
-                    return this.transfer(toolActions.changeVectorMaskMode, false);
+                    return this.transfer("tools.changeVectorMaskMode", false);
                 });
         }
 
         var currentLayer = currentDocument.layers.selected.first();
 
-        return this.transfer(layerActions.resetLayers, currentDocument, currentLayer)
+        return this.transfer("layers.resetLayers", currentDocument, currentLayer)
             .bind(this)
             .then(function () {
                 currentDocument = appStore.getCurrentDocument();
@@ -118,27 +208,27 @@ define(function (require, exports) {
                 
                 if (currentLayer && !currentLayer.vectorMaskEmpty) {
                     var payload = { commandID: _CLEAR_PATH, waitForCompletion: true };
-                    return this.transfer(menuActions.native, payload)
+                    return this.transfer("menu.native", payload)
                         .bind(this)
                         .then(function () {
-                            return this.transfer(layerActions.resetLayers, currentDocument, currentLayer);
+                            return this.transfer("layers.resetLayers", currentDocument, currentLayer);
                         })
                         .then(function () {
                             currentDocument = appStore.getCurrentDocument();
                             currentLayer = currentDocument.layers.selected.first();
                             if (currentLayer && !currentLayer.vectorMaskEnabled) {
-                                return this.transfer(layerActions.deleteVectorMask)
+                                return this.transfer("mask.deleteVectorMask")
                                     .bind(this)
                                     .then(function () {
-                                        return this.transfer(toolActions.changeVectorMaskMode, false);
+                                        return this.transfer("tools.changeVectorMaskMode", false);
                                     });
                             }
                         });
                 } else {
-                    return this.transfer(layerActions.deleteVectorMask)
+                    return this.transfer("mask.deleteVectorMask")
                         .bind(this)
                         .then(function () {
-                            this.transfer(toolActions.changeVectorMaskMode, false);
+                            this.transfer("tools.changeVectorMaskMode", false);
                         });
                 }
             });
@@ -146,12 +236,12 @@ define(function (require, exports) {
     handleDeleteVectorMask.action = {
         reads: [locks.JS_APP, locks.JS_DOC],
         writes: [],
-        transfers: [layerActions.resetLayers, layerActions.deleteVectorMask, toolActions.changeVectorMaskMode,
-            menuActions.native]
+        transfers: ["layers.resetLayers", "mask.deleteVectorMask", "tools.changeVectorMaskMode",
+            "menu.native"]
     };
 
     /**
-     * Create an mask that matches the bounds of the currently selected layer
+     * Create a mask that matches the bounds of the currently selected layer
      *
      * @return {Promise}
      */
@@ -186,17 +276,17 @@ define(function (require, exports) {
                 vectorMaskLib.deleteWorkPath()], options)
             .bind(this)
             .then(function () {
-                return this.transfer(toolActions.select, toolStore.getToolByID("newSelect"));
+                return this.transfer("tools.select", toolStore.getToolByID("newSelect"));
             });
     };
     applyRectangle.action = {
         reads: [locks.JS_APP, locks.JS_DOC],
         writes: [locks.PS_DOC],
-        transfers: [toolActions.select]
+        transfers: ["tools.select"]
     };
 
     /**
-     * Create an circular mask the size of the currently selected layer
+     * Create a circular mask the size of the currently selected layer
      *
      * @return {Promise}
      */
@@ -231,15 +321,18 @@ define(function (require, exports) {
                 vectorMaskLib.deleteWorkPath()], options)
             .bind(this)
             .then(function () {
-                return this.transfer(toolActions.select, toolStore.getToolByID("newSelect"));
+                return this.transfer("tools.select", toolStore.getToolByID("newSelect"));
             });
     };
     applyEllipse.action = {
         reads: [locks.JS_APP, locks.JS_DOC],
         writes: [locks.PS_DOC],
-        transfers: [toolActions.select]
+        transfers: ["tools.select"]
     };
 
+    exports.editVectorMask = editVectorMask;
+    exports.selectVectorMask = selectVectorMask;
+    exports.deleteVectorMask = deleteVectorMask;
     exports.handleDeleteVectorMask = handleDeleteVectorMask;
     exports.handleEscapeVectorMask = handleEscapeVectorMask;
     exports.applyEllipse = applyEllipse;
